Split mock instances to gamechannel/testutils.
Split up the gamechannel/testgame file, moving the mock classes to
a new testutils file that is also exposed as a libtool library, so
that e.g. the ships tests can use the mock instances.
